pyproject.toml
[project]
name = "insider"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.30.0",
]

[project.scripts]
insider = "insider.cli:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
